chore(core): move `paired_cache` to `trezor.wire.thp`
Also, log `ThpPairedCache`-related events.

[no changelog]
