+++ core/embed/upymod/qstrdefsport.h
@@ -407,7 +407,6 @@ Q(ThpPairingMethod)
 Q(alternating_bit_protocol)
 Q(apps.thp)
 Q(apps.thp.credential_manager)
-Q(apps.thp.paired_cache)
 Q(apps.thp.pairing)
 Q(cache_thp)
 Q(cache_thp_keys)
@@ -441,6 +440,7 @@ Q(trezor.wire.thp.cpace)
 Q(trezor.wire.thp.crypto)
 Q(trezor.wire.thp.interface_context)
 Q(trezor.wire.thp.memory_manager)
+Q(trezor.wire.thp.paired_cache)
 Q(trezor.wire.thp.pairing_context)
 Q(trezor.wire.thp.received_message_handler)
 Q(trezor.wire.thp.session_context)
